+++ CHANGES.rst
@@ -11,6 +11,8 @@ Features added
 * Support for coverage.py 4.0+ can be enabled by adding the plugin
   "Cython.Coverage" to the ".coveragerc" config file.
 
+* Tracing is supported in ``nogil`` functions/sections.
+
 Bugs fixed
 ----------